requirements.txt
asgiref==3.7.2
Django==4.2.1
gunicorn==20.1.0
Pillow==9.5.0
sqlparse==0.4.4
typing_extensions==4.6.2
whitenoise==6.4.0
